refactor: integrate handler tests
